Add in implemetation for furnace macros

+++ src/macros.h
@@ -14,27 +14,27 @@ typedef enum {
 void clearmacros(void);
 
 // Setup sequence macro
-void sequencemacro(uint16_t ID, const int *sequence, uint8_t length, uint8_t looppoint, uint8_t releasepoint);
+void sequencemacro(uint8_t ID, const int *sequence, uint8_t length, uint8_t looppoint, uint8_t releasepoint);
 
 // Setup ADSR macro
-void adsrmacro(uint16_t ID, int top, int bottom,
+void adsrmacro(uint8_t ID, int top, int bottom,
                uint8_t atk, uint8_t hld, uint8_t dcy,
                uint8_t sus, uint8_t sus_t, uint8_t sus_d, uint8_t rel);
 
 // Setup LFO macro
-void lfomacro(uint16_t ID, int top, int bottom, uint8_t speed, lfowave wave, uint16_t phase);
+void lfomacro(uint8_t ID, int top, int bottom, uint8_t speed, lfowave wave, uint16_t phase);
 
 // Set global parameters for a macro
-void globalparams(uint16_t ID, uint8_t delay, uint8_t step);
+void globalparams(uint8_t ID, uint8_t delay, uint8_t step);
 
 // Start macro with given ID and attach value pointer
-void startmacro(uint16_t ID, int* value);
+void startmacro(uint8_t ID, int* value);
 
 // Stop macro with given ID
-void stopmacro(uint16_t ID);
+void stopmacro(uint8_t ID);
 
 // Release macro with given ID
-void releasemacro(uint16_t ID);
+void releasemacro(uint8_t ID);
 
 // Process all active macros per tick
 void tickmacros(void);
